feat: Seperate class files
Seperated out class files for Recurrence/Rules/Pattern within domain models
