+++ service/grails-app/domain/org/olf/recurrence/recurrencePattern/RecurrencePattern.groovy
@@ -1,0 +1,8 @@
+package org.olf
+
+import grails.gorm.MultiTenant
+
+public class RecurrencePattern implements MultiTenant<RecurrencePattern> {
+  String id
+  RecurrenceRule owner
+}
