+++ service/grails-app/domain/org/olf/recurrence/recurrencePattern/RecurrencePatternDay.groovy
@@ -1,0 +1,7 @@
+package org.olf
+
+import grails.gorm.MultiTenant
+
+public class RecurrencePatternDay extends RecurrencePattern implements MultiTenant<RecurrencePatternDay> {
+
+}
