+++ service/grails-app/domain/org/olf/recurrence/recurrencePattern/RecurrencePatternMonthDate.groovy
@@ -1,0 +1,8 @@
+package org.olf
+
+import grails.gorm.MultiTenant
+
+public class RecurrencePatternMonthDate extends RecurrencePattern implements MultiTenant<RecurrencePatternMonthDate> {
+  Integer day // Validated 1-28 and -1 // OR special refdata
+  // Fallback stuff?
+}
